Initial Attempt at Noncanonical FFTS
Initial try to see if we can easily implement delayed reduction for FFT's.

Had to shuffle around code to prevent a circular dependancy chain.
