+++ baby-bear/src/lib.rs
@@ -4,8 +4,10 @@ extern crate alloc;
 
 mod baby_bear;
 mod extension;
+mod non_canonical;
 
 pub use baby_bear::*;
+pub use non_canonical::*;
 
 #[cfg(all(target_arch = "aarch64", target_feature = "neon"))]
 mod aarch64_neon;
