+++ field/src/lib.rs
@@ -8,6 +8,7 @@ mod array;
 mod batch_inverse;
 mod exponentiation;
 pub mod extension;
+mod non_canonical;
 mod field;
 mod helpers;
 mod packed;
@@ -20,3 +21,4 @@ pub use field::*;
 pub use helpers::*;
 pub use packed::*;
 pub use symbolic::*;
+pub use non_canonical::*;
