+++ mds/src/karatsuba_convolution.rs
@@ -1,10 +1,6 @@
-use alloc::vec;
-use alloc::vec::Vec;
-use core::ops::{Add, AddAssign, Mul, MulAssign, Shl, ShlAssign, Shr, ShrAssign, Sub, SubAssign};
 // use core::mem::transmute;
 
-use crate::types::{Canonicalize, IntegerLike, NonCanonicalPrimeField32};
-use p3_field::{PrimeField32, PrimeField64};
+use p3_field::{PrimeField32, Canonicalize, IntegerLike, NonCanonicalPrimeField32};
 
 /// Computes the convolution of input and MATRIX_CIRC_MDS_8_SML.
 /// Input must be an array of field elements of length 8.
@@ -868,253 +864,3 @@ impl Convolution for SmallConvolution {
 // Would be good to try and find a way to cut down on this.
 
 // Once we get down to small sizes we use the O(n^2) approach.
-
-///////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-// Generic functions
-
-///////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-// Assume that lhs is a vector of u64's and rhs is a vector of field elements.
-// This computes M(lhs).rhs where M is the circulant matrix with first row equal to lhs.
-// It uses the even-odd decomposition.
-/// Currently have not implmented the FFT trick to use signed convolutions.
-pub fn apply_circulant_karat_generic_i64<F: PrimeField64, const N: usize>(
-    lhs: [F; N],
-    rhs: [i64; N],
-) -> [F; N] {
-    // Convert lhs from field elements to i64's
-    let lhs_i64 = lhs.map(|x| x.as_canonical_u64() as i64);
-
-    // We need the vector which is the first column of rhs not the first row.
-    let rhs_i64 = row_to_col(rhs);
-
-    let mut output = [0; N];
-
-    conv_karat_generic(&lhs_i64, &rhs_i64, &mut output);
-
-    output.map(|x| F::from_wrapped_u64(x as u64))
-}
-
-// It will be handy for functions to be able to handle entries which are a combination of simple integer types
-// In particular u64's, i64's, u128's and i128's so we make a general trait type here.
-trait SimpleInteger:
-    Sized
-    + Default
-    + Copy
-    + Add<Output = Self>
-    + AddAssign
-    + Sub<Output = Self>
-    + SubAssign
-    + Mul<Output = Self>
-    + MulAssign
-    + ShrAssign<usize>
-    + Shr<usize, Output = Self>
-    + ShlAssign<usize>
-    + Shl<usize, Output = Self>
-{
-}
-
-impl<T> SimpleInteger for T where
-    T: Sized
-        + Default
-        + Copy
-        + Add<Output = Self>
-        + AddAssign
-        + Sub<Output = Self>
-        + SubAssign
-        + Mul<Output = Self>
-        + MulAssign
-        + ShrAssign<usize>
-        + Shr<usize, Output = Self>
-        + ShlAssign<usize>
-        + Shl<usize, Output = Self>
-{
-}
-
-impl<T> IntegerLike for T where T: SimpleInteger {}
-
-// Let M be a circulant matrix with first column vec_col and first row vec_row. Then M.u is the convolution of vec_col and u.
-// The vectors given here are the first rows of the respective circulant matrices NOT the first colums.
-// Hence in order to do convolutions we need to compute the first column which is given by
-// vec_col = [vec_row[0], vec_row[n - 1], vec_row[n - 2], ..., vec_row[1]]
-#[inline]
-const fn row_to_col<T: SimpleInteger, const N: usize>(row: [T; N]) -> [T; N] {
-    let mut col = [row[0]; N];
-    let mut i = 1;
-    loop {
-        if i == N {
-            break;
-        }
-        col[i] = row[N - i];
-        i += 1
-    }
-    col
-}
-
-// Takes the dot product of two vectors.
-#[inline]
-fn dot<T: SimpleInteger>(lhs: &[T], rhs: &[T]) -> T {
-    let n = lhs.len();
-    let mut sum = lhs[0] * rhs[0];
-    for i in 1..n {
-        sum += lhs[i] * rhs[i];
-    }
-    sum
-}
-
-// We produce a generic implementations. It will likely be faster long term to specialise these.
-// Given lhs (v) and rhs (u) compute the convolution of lhs and rhs recursively as smaller convolutions and signed convolutions.
-// Currently this only works for n a power of 2.
-fn conv_karat_generic<T: SimpleInteger>(lhs: &[T], rhs: &[T], output: &mut [T]) {
-    let n = lhs.len();
-    match n {
-        1 => output[0] = lhs[0] * rhs[0],
-        2 => {
-            output[0] = lhs[0] * rhs[0] + lhs[1] * rhs[1];
-            output[1] = lhs[1] * rhs[0] + lhs[0] * rhs[1];
-        }
-        4 => conv4_slice(lhs, rhs, output),
-        _ => {
-            let half = n / 2;
-
-            let (lhs_left, lhs_right) = lhs.split_at(half);
-
-            let mut lhs_p = lhs_left.to_vec();
-            let mut lhs_m = lhs_left.to_vec();
-
-            add_mut(&mut lhs_p, lhs_right); // lhs(x) mod x^{n/2} - 1
-            sub_mut(&mut lhs_m, lhs_right); // lhs(x) mod x^{n/2} + 1
-
-            // rhs will always be constant. Not sure how to tell the compiler this though.
-            let (rhs_left, rhs_right) = rhs.split_at(half);
-
-            let mut rhs_p = rhs_left.to_vec();
-            let mut rhs_m = rhs_left.to_vec();
-
-            add_mut(&mut rhs_p, rhs_right); // rhs(x) mod x^{n/2} - 1
-            sub_mut(&mut rhs_m, rhs_right); // rhs(x) mod x^{n/2} + 1
-
-            let (left, right) = output.split_at_mut(half);
-            signed_conv_karat_generic(&lhs_m, &rhs_m, left); // left = w_1 = lhs*rhs mod x^32 + 1
-            conv_karat_generic(&lhs_p, &rhs_p, right); // right = w_0 = lhs*rhs mod x^32 - 1
-            for i in 0..half {
-                left[i] += right[i]; // w_0 + w_1
-                left[i] >>= 1; // (w_0 + w_1)/2
-                right[i] -= left[i]; // (w_0 - w_1)/2
-            }
-        }
-    }
-}
-
-/// Compute the convolution of two vectors of length 4.
-/// output(x) = lhs(x)rhs(x) mod x^4 - 1
-#[inline]
-fn conv4_slice<T: SimpleInteger>(lhs: &[T], rhs: &[T], output: &mut [T]) {
-    // lhs.len(), rhs.len() and output.len() should all be equal to 4.
-    // Add constant asserts?
-    // Even at this small size, doing the FFT decomposition seems to produce shorter compiled code using godbolt.
-
-    let lhs_p = [lhs[0] + lhs[2], lhs[1] + lhs[3]]; // v_0(x)
-    let lhs_m = [lhs[0] - lhs[2], lhs[1] - lhs[3]]; // v_1(x)
-
-    // In most cases, rhs will actually be a constant but I'm unsure how to take advantage of this.
-    let rhs_p = [rhs[0] + rhs[2], rhs[1] + rhs[3]]; // u_0(x)
-    let rhs_m = [rhs[0] - rhs[2], rhs[1] - rhs[3]]; // u_1(x)
-
-    // Might be worth trying to keep everything as a i64 up until this multiplication and
-    // only here making things u128's. (Possible that the compiler has already worked this out though.)
-    output[0] = lhs_m[0] * rhs_m[0] - lhs_m[1] * rhs_m[1];
-    output[1] = lhs_m[0] * rhs_m[1] + lhs_m[1] * rhs_m[0]; // output[0, 1] = w_1 = v_1(x)u_1(x) mod x^2 + 1
-    output[2] = lhs_p[0] * rhs_p[0] + lhs_p[1] * rhs_p[1];
-    output[3] = lhs_p[0] * rhs_p[1] + lhs_p[1] * rhs_p[0]; // output[2, 3] = w_0 = v_0(x)u_0(x) mod x^2 - 1
-
-    output[0] += output[2];
-    output[1] += output[3]; // output[0, 1] = w_1 + w_0
-
-    output[0] >>= 1;
-    output[1] >>= 1; // output[0, 1] = (w_1 + w_0)/2)
-
-    output[2] -= output[0];
-    output[3] -= output[1]; // output[2, 3] = w_0 - (w_1 + w_0)/2) = (w_0 - w_1)/2
-}
-
-/// Compute the signed convolution of two vectors of length 4.
-/// output(x) = lhs(x)rhs(x) mod x^4 + 1
-/// Same as above but in place.
-#[inline]
-fn signed_conv4_slice<T: SimpleInteger>(lhs: &[T], rhs: &[T], output: &mut [T]) {
-    // lhs.len(), rhs.len() and output.len() should all be equal to 4.
-    let rhs_rev = [rhs[3], rhs[2], rhs[1], rhs[0]];
-
-    output[0] = lhs[0] * rhs[0] - dot(&lhs[1..], &rhs_rev[..3]); // v_0u_0 - (v_1u_3 + v_2u_2 + v_3u_1)
-    output[1] = dot(&lhs[..2], &rhs_rev[2..]) - dot(&lhs[2..], &rhs_rev[..2]); // v_0u_1 + v_1u_0 - (v_2u_3 + v_2u_3)
-    output[2] = dot(&lhs[..3], &rhs_rev[1..]) - lhs[3] * rhs[3]; // v_0u_2 + v_1u_1 + v_2u_0 - v_3u_3
-    output[3] = dot(lhs, &rhs_rev); // v_0u_3 + v_1u_2 + v_2u_1 + v_3u_0
-}
-
-// Given lhs (v) and rhs (u) compute the signed convolution via the karatsuba method.
-fn signed_conv_karat_generic<T: SimpleInteger>(lhs: &[T], rhs: &[T], output: &mut [T]) {
-    let n = lhs.len();
-    match n {
-        1 => output[0] = lhs[0] * rhs[0],
-        2 => {
-            output[0] = lhs[0] * rhs[0] - lhs[1] * rhs[1];
-            output[1] = lhs[1] * rhs[0] + lhs[0] * rhs[1];
-        }
-        4 => signed_conv4_slice(lhs, rhs, output),
-        _ => {
-            let half = n / 2;
-            let (lhs_even, lhs_odd, lhs_mix) = split_eom(lhs);
-            let (rhs_even, rhs_odd, rhs_mix) = split_eom(rhs);
-
-            {
-                let (evens, odds) = output.split_at_mut(half);
-                let mut extra = vec![T::default(); half];
-
-                signed_conv_karat_generic(&lhs_even, &rhs_even, evens); // v_e(x)u_e(x) mod x^{n/2} + 1
-                signed_conv_karat_generic(&lhs_odd, &rhs_odd, &mut extra); // v_o(x)u_o(x) mod x^{n/2} + 1
-                signed_conv_karat_generic(&lhs_mix, &rhs_mix, odds); // (v_e(x) + v_o(x))(u_e(x) + u_o(x)) mod x^{n/2} + 1
-
-                sub_mut(odds, evens);
-                sub_mut(odds, &extra); // (v_e(x) + v_o(x))(u_e(x) + u_o(x)) - v_e(x)u_e(x) - v_o(x)u_o(x)
-
-                add_mut(&mut evens[1..], &extra[..(half - 1)]);
-                evens[0] -= extra[half - 1]; // v_e(x)u_e(x) + xv_o(x)u_o(x) mod x^16 + 1
-            }
-
-            rearrange(output);
-        }
-    }
-}
-
-/// Given a vector v, split it into its even, odd and mixed parts.
-/// If vec = [v_0, v_1, v_2, ...] then
-/// output = ([v_0, v_2, ...], [v_1, v_3, ...], [v_0 + v_1, v_2 + v_3, ...])
-#[inline]
-fn split_eom<T: SimpleInteger>(vec: &[T]) -> (Vec<T>, Vec<T>, Vec<T>) {
-    let half = vec.len() / 2;
-    let mut output_even = vec![T::default(); half];
-    let mut output_odd = vec![T::default(); half];
-    let mut output_mix = vec![T::default(); half];
-    for i in 0..half {
-        output_even[i] = vec[2 * i];
-        output_odd[i] = vec[2 * i + 1];
-        output_mix[i] = vec[2 * i] + vec[2 * i + 1];
-    }
-    (output_even, output_odd, output_mix)
-}
-
-/// Given a vector v, interleave it's left and right parts in place
-/// if initially vec = [v_0, v_1, ..., v_N, v_{N + 1}, ..., V_{2N - 1}]
-/// afterwards vec = [v_0, v_N, v_1, v_{N + 1}, ..., v_{N - 1}, v_{2N - 1}]
-#[inline]
-fn rearrange<T: SimpleInteger>(vec: &mut [T]) {
-    let n = vec.len();
-    let half = n / 2;
-    let stored_evens = vec[..half].to_vec().clone();
-    for i in 0..half {
-        vec[2 * i] = stored_evens[i];
-        vec[2 * i + 1] = vec[i + half];
-    }
-}
