+++ mds/src/lib.rs
@@ -13,7 +13,6 @@ pub mod goldilocks;
 pub mod integrated_coset_mds;
 pub mod karatsuba_convolution;
 pub mod mersenne31;
-mod types;
 pub mod util;
 
 pub trait MdsPermutation<T: Clone, const WIDTH: usize>: Permutation<[T; WIDTH]> {}
